qiskit-code-assistant-mcp-server/pyproject.toml
[project]
name = "qiskit-code-assistant-mcp-server"
version = "0.2.0"
description = "MCP server for the Qiskit Code Assistant"
readme = "README.md"
requires-python = ">=3.10,<3.15"
dependencies = [
    "fastmcp>=2.8.1",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "nest-asyncio>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.11.13",
    "mypy>=1.15.0",
    "pre-commit>=3.5.0",
    "bandit[toml]>=1.7.0",
]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
    "respx>=0.20.0",
]
all = [
    "qiskit-code-assistant-mcp-server[dev,test]",
]
[[project.authors]]
name = "Quantum+AI Team. IBM Quantum"
email = "Quantum.Plus.AI@ibm.com"

[project.license]
text = "Apache-2.0"

keywords = ["quantum", "qiskit", "mcp", "ai", "code-assistant", "quantum-computing", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.urls]
Homepage = "https://github.com/Qiskit/mcp-servers"
Repository = "https://github.com/Qiskit/mcp-servers"
Documentation = "https://github.com/Qiskit/mcp-servers/tree/main/qiskit-code-assistant-mcp-server#readme"
"Bug Tracker" = "https://github.com/Qiskit/mcp-servers/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/qiskit_code_assistant_mcp_server"]

[project.scripts]
qiskit-code-assistant-mcp-server = "qiskit_code_assistant_mcp_server:main"

[dependency-groups]
dev = [
    "ruff>=0.11.13",
    "mypy>=1.15.0",
    "bandit[toml]>=1.9.2",
]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
    "respx>=0.20.0",
]
examples = [
    "langchain>=1.2.0",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-anthropic>=1.0.0",
    "python-dotenv>=1.0.0",
]

# Bandit security configuration
[tool.bandit]
skips = ["B101"]  # Skip assert_used check in tests
exclude_dirs = ["tests", ".venv", "build", "dist"]
